requirements.txt
Flask==3.0.2
flask-login==0.6.3
python-dotenv==1.0.1
openai
google-generativeai==0.3.2
chromadb==0.4.18
langchain-openai>=0.0.7
langchain-core>=0.1.16,<0.2.0
psycopg2-binary
python-docx==1.0.1
beautifulsoup4==4.12.3
Pillow>=9.5.0
pytesseract==0.3.10
pdf2image==1.17.0
selenium==4.18.1
webdriver-manager==4.0.1
Werkzeug==3.0.1
gunicorn==21.2.0
protobuf>=4.25.3,<5.0.0
